Testing move control plane components
This moves control plane components to a new internal/testing/controlplane
directory, unifying the first and second levels of internal packages.
