+++ pkg/internal/testing/controlplane/kubectl.go
@@ -1,4 +1,4 @@
-package integration
+package controlplane
 
 import (
 	"bytes"
