+++ pkg/internal/testing/controlplane/plane.go
@@ -1,4 +1,4 @@
-package integration
+package controlplane
 
 import (
 	"fmt"
